move and fix forms for user registration
