--- pages/forms.py
@@ -1,13 +1,0 @@
-from api import settings
-from django import forms
-from django.contrib.auth import get_user_model
-User = get_user_model()
-from django.contrib.auth.forms import UserCreationForm
-
-class CustomUserCreationForm(UserCreationForm):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.fields['email'].required = True
-    class Meta(UserCreationForm.Meta):
-        model = User
-        fields = UserCreationForm.Meta.fields + ('first_name', 'last_name', 'email')
